Get Auth in angular app working

+++ app/Http/Controllers/PrizeController.php
@@ -5,28 +5,38 @@ namespace App\Http\Controllers;
 use App\Prize;
 use Illuminate\Http\Request;
 use Exception;
+use Illuminate\Support\Facades\Auth;
 
 class PrizeController extends Controller
 {
     public function index() {
-      $prizes = Prize::all();
-      return response()->json($prizes);
+        if(!Auth::check()) {
+            return response()->json('Denied', 401);
+        }
+        $prizes = Prize::all();
+        return response()->json($prizes);
     }
 
     public function enabled() {
-      $enabledPrizes = Prize::where('enabled', 1)->get();
-      return response()->json($enabledPrizes);
+        if(!Auth::check()) {
+            return response()->json('Denied', 401);
+        }
+        $enabledPrizes = Prize::where('enabled', 1)->get();
+        return response()->json($enabledPrizes);
     }
 
     public function create(Request $request) {
+        if(!Auth::check()) {
+            return response()->json('Denied', 401);
+        }
 
         $validator = \Validator::make($request->all(), [
-          'name' => 'required',
-          'enabled' => 'boolean'
+            'name' => 'required',
+            'enabled' => 'boolean'
         ]);
 
         if ($validator->fails()) {
-          return response()->json($validator);
+            return response()->json($validator);
         }
 
         $prize = Prize::create($request->all());
@@ -35,6 +45,9 @@ class PrizeController extends Controller
     }
 
     public function update(Request $request, Prize $prize) {
+        if(!Auth::check()) {
+            return response()->json('Denied', 401);
+        }
         if($request->has('name')) {
             $prize->name = $request->name;
         }
@@ -49,6 +62,9 @@ class PrizeController extends Controller
     }
 
     public function destroy(Request $request, Prize $prize) {
+        if(!Auth::check()) {
+            return response()->json('Denied', 401);
+        }
         try {
             $prize->delete();
             return response()->json('deleted');
